generated_projects/mailanalyser_4cc01bae/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
imaplib2==2.57.3
email
WTForms==3.0.1
flask-wtf==1.1.1